requirements.txt
flask==2.2.4
requests
PyJWT
waitress
flask_talisman
gunicorn


